refactor: 무한 스크롤을 담당하는 훅 분리
[#412]
